Added exception handling and updating unit tests

+++ src/main/java/net/kreddo/notes/controller/NoteController.java
@@ -3,9 +3,13 @@ package net.kreddo.notes.controller;
 import java.util.List;
 import javax.validation.Valid;
 import lombok.RequiredArgsConstructor;
+import lombok.extern.slf4j.Slf4j;
 import net.kreddo.notes.controller.dto.NoteDto;
+import net.kreddo.notes.controller.exception.InvalidInputException;
 import net.kreddo.notes.service.NoteService;
+import org.springframework.http.HttpStatus;
 import org.springframework.http.MediaType;
+import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PatchMapping;
@@ -15,6 +19,7 @@ import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 
+@Slf4j
 @RestController
 @RequiredArgsConstructor
 @RequestMapping("/notes")
@@ -28,28 +33,48 @@ public class NoteController {
   }
 
   @GetMapping(path = "/{id}", produces = MediaType.APPLICATION_JSON_VALUE)
-  public NoteDto getNote(@PathVariable("id") Long id) {
-    return noteService.getNote(id);
+  public ResponseEntity<NoteDto> getNote(@PathVariable("id") Long id) {
+    if (null == id || id.equals(0L)) {
+      throw new InvalidInputException("Invalid Note id");
+    }
+    NoteDto note = noteService.getNote(id);
+    return new ResponseEntity<>(note, HttpStatus.OK);
   }
 
   @PostMapping(value = "/", consumes = MediaType.APPLICATION_JSON_VALUE)
-  public NoteDto addNote(@RequestBody @Valid NoteDto noteDto) {
-    return noteService.addNote(noteDto);
+  public ResponseEntity<NoteDto> addNote(@RequestBody @Valid NoteDto noteDto) {
+    if (noteDto.getId() != null) {
+      throw new InvalidInputException("Note Id should not exist");
+    }
+    NoteDto note = noteService.addNote(noteDto);
+    return new ResponseEntity<>(note, HttpStatus.CREATED);
   }
 
   @PatchMapping(value = "/{id}", consumes = MediaType.APPLICATION_JSON_VALUE)
-  public NoteDto updateNote(@RequestBody @Valid NoteDto noteDto) {
-    return noteService.updateNote(noteDto);
+  public ResponseEntity<NoteDto> updateNote(@RequestBody @Valid NoteDto noteDto) {
+    if (noteDto.getId() == null) {
+      throw new InvalidInputException("Note Id should exist");
+    }
+    NoteDto note = noteService.updateNote(noteDto);
+    return new ResponseEntity<>(note, HttpStatus.OK);
   }
 
   @DeleteMapping("/{id}")
-  public void deleteNote(@PathVariable("id") Long id) {
+  public ResponseEntity<?> deleteNote(@PathVariable("id") Long id) {
+    if (null == id || id.equals(0L)) {
+      throw new InvalidInputException("Invalid Note id");
+    }
     noteService.deleteNote(id);
+    return new ResponseEntity<>(HttpStatus.OK);
   }
 
   @DeleteMapping("/note")
-  public void deleteNote(@RequestBody @Valid NoteDto noteDto) {
+  public ResponseEntity<?> deleteNote(@RequestBody @Valid NoteDto noteDto) {
+    if (noteDto.getId() == null) {
+      throw new InvalidInputException("Note Id should exist");
+    }
     noteService.deleteNote(noteDto);
+    return new ResponseEntity<>(HttpStatus.OK);
   }
 
 }
